Part: Extension of Geometry::Curve to wrap more functions of OCC
Apart from:
    bool tangent(double u, gp_Dir&) const;
that was already implemented, now it also implements:
    Base::Vector3d pointAtParameter(double u) const;
    Base::Vector3d firstDerivativeAtParameter(double u) const;
    Base::Vector3d secondDerivativeAtParameter(double u) const;
    Base::Vector3d thirdDerivativeAtParameter(double u) const;
    bool normal(double u, gp_Dir& dir) const;
    bool closestParameter(Base::Vector3d point, double &u);
    bool closestParameterToBasicCurve(Base::Vector3d point, double &u);

i.e. apart from giving the tangent vector for a given curve at parameter value u,
it also gives:
   - The point cartesian coordinates of the curve point at parameter value u
   - The vector of the first derivative at parameter value u
   - The vector of the second derivative at parameter value u
   - The normal vector to the curve at parameter value u
   - The parameter value of a curve closest ot a given point
   - The same as the one before, but using the basic curve of a trimmedcurve
(useful e.g. to get the point closest to the ellipse used to define an arc of ellipse).
It provides the parameter of the curve closest to a given point.
If the curve is a Trimmed curve, the parameter of the basic underlaying curve closest
to the point is provided.

NOTE: Tangent and Normal require 2 derivatives, because the OCC algorithm uses the value of
the second derivative in case the first is zero for tangent and because normal uses
curvature in the calculus which requries 2 derivatives.